fix: resolve all clippy diagnostics — approx_constant, manual_div_ceil, needless_range_loop, unused_mut, missing cfg gate
- src/octopus/encode.rs: 0.7071 → FRAC_1_SQRT_2
- src/octopus/octahedral.rs: manual div_ceil → .div_ceil()
- src/hybrid_oct_pq/kv_cache.rs: remove unnecessary as f32 casts
- src/planar_quant/rotation.rs: div_ceil + let_and_return fixes
- src/iso_quant/rotation.rs: div_ceil + approx_constant fixes
- tests/bench_block_diagonal_goat.rs: add #[cfg(feature = "maxsim")] gate, #[allow(unused_mut)]
- tests/bench_octopus_goat.rs: remove unnecessary cast + div_ceil

+++ src/hybrid_oct_pq/kv_cache.rs
@@ -412,7 +412,7 @@ mod tests {
 
     fn make_random_vec(dim: usize, seed: u64) -> Vec<f32> {
         let mut rng = crate::types::Rng::new(seed);
-        (0..dim).map(|_| rng.normal() as f32).collect()
+        (0..dim).map(|_| rng.normal()).collect()
     }
 
     fn cosine_sim(a: &[f32], b: &[f32]) -> f32 {
@@ -744,7 +744,7 @@ mod tests {
 
         let mut rng = crate::types::Rng::new(seed);
         let keys: Vec<Vec<f32>> = (0..n_keys)
-            .map(|_| (0..dim).map(|_| rng.normal() as f32).collect())
+            .map(|_| (0..dim).map(|_| rng.normal()).collect())
             .collect();
 
         // Hybrid MSE
